requirement.txt
Package             Version
------------------- --------
anyio               4.4.0
blinker             1.8.2
certifi             2024.7.4
click               8.1.7
exceptiongroup      1.2.2
Flask               3.0.3
h11                 0.14.0
httpcore            1.0.5
httpx               0.27.0
idna                3.7
itsdangerous        2.2.0
Jinja2              3.1.4
MarkupSafe          2.1.5
pip                 22.0.2
pyaes               1.6.1
pyasn1              0.6.0
python-dotenv       1.0.1
python-telegram-bot 21.4
rsa                 4.9
secure-smtplib      0.1.1
setuptools          59.6.0
sniffio             1.3.1
Telethon            1.36.0
typing_extensions   4.12.2
Werkzeug            3.0.3
